+++ demo/oauth2-github.r
@@ -15,6 +15,12 @@ myapp <- oauth_app("github", "56b637a5baffac62cad9")
 github_token <- oauth2.0_token(oauth_endpoints("github"), myapp)
 
 # 4. Use API
-req <- GET("https://api.github.com/rate_limit", config(token = github_token))
+gtoken <- config(token = github_token)
+req <- GET("https://api.github.com/rate_limit", gtoken)
+stop_for_status(req)
+content(req)
+
+# OR:
+req <- with_config(gtoken, GET("https://api.github.com/rate_limit"))
 stop_for_status(req)
 content(req)